test: update shared view test cases
