+++ packages/nc-gui/components/shared-view/AskPassword.vue
@@ -104,7 +104,15 @@ const bgImageName = computed(() => {
         </a-form-item>
       </a-form>
       <div class="flex flex-row justify-end gap-x-2">
-        <NcButton :disabled="!formState.password" type="primary" size="small" html-type="submit" class="!px-2" @click="onFinish">
+        <NcButton
+          :disabled="!formState.password"
+          type="primary"
+          size="small"
+          html-type="submit"
+          class="!px-2"
+          data-testid="nc-shared-view-password-submit-btn"
+          @click="onFinish"
+        >
           {{ $t('objects.view') }}
           <template #loading> {{ $t('msg.verifyingPassword') }}</template>
         </NcButton>
