+++ tests/playwright/pages/Dashboard/common/Toolbar/index.ts
@@ -195,28 +195,6 @@ export class ToolbarPage extends BasePage {
     await this.get().locator(`.nc-toolbar-btn.nc-add-new-row-btn`).click();
   }
 
-  async clickDownload(type: string, verificationFile = 'expectedData.txt') {
-    await this.get().locator(`.nc-toolbar-btn.nc-actions-menu-btn`).click();
-
-    const [download] = await Promise.all([
-      // Start waiting for the download
-      this.rootPage.waitForEvent('download'),
-      // Perform the action that initiates download
-      this.rootPage
-        .locator(`.nc-dropdown-actions-menu`)
-        .locator(`li.ant-dropdown-menu-item:has-text("${type}")`)
-        .click(),
-    ]);
-
-    // Save downloaded file somewhere
-    await download.saveAs('./output/at.txt');
-
-    // verify downloaded content against expected content
-    const expectedData = fs.readFileSync(`./fixtures/${verificationFile}`, 'utf8').replace(/\r/g, '').split('\n');
-    const file = fs.readFileSync('./output/at.txt', 'utf8').replace(/\r/g, '').split('\n');
-    expect(file).toEqual(expectedData);
-  }
-
   async clickRowHeight() {
     // ant-btn nc-height-menu-btn nc-toolbar-btn
     await this.get().locator(`.nc-toolbar-btn.nc-height-menu-btn`).click();
