+++ tests/playwright/tests/db/views/viewGridShare.spec.ts
@@ -246,7 +246,7 @@ test.describe('Shared view', () => {
      **/
 
     // verify download
-    await sharedPage.grid.toolbar.clickDownload(
+    await sharedPage.grid.topbar.clickDownload(
       'Download CSV',
       isSqlite(context) || isPg(context) ? 'expectedDataSqlite.txt' : 'expectedData.txt'
     );
@@ -284,12 +284,12 @@ test.describe('Shared view', () => {
     // verify if password request modal exists
     const sharedPage2 = new DashboardPage(page, context.base);
     await sharedPage2.rootPage.locator('input[placeholder="Enter password"]').fill('incorrect p@ssword');
-    await sharedPage2.rootPage.click('button:has-text("Unlock")');
+    await sharedPage2.rootPage.click('button[data-testid="nc-shared-view-password-submit-btn"]');
     await sharedPage2.verifyToast({ message: 'INVALID_SHARED_VIEW_PASSWORD' });
 
     // correct password
     await sharedPage2.rootPage.locator('input[placeholder="Enter password"]').fill('p@ssword');
-    await sharedPage2.rootPage.click('button:has-text("Unlock")');
+    await sharedPage2.rootPage.click('button[data-testid="nc-shared-view-password-submit-btn"]');
 
     // verify if download button is disabled
     await sharedPage2.grid.toolbar.verifyDownloadDisabled();
